Fix issue with logging to stdout
Using the constant `STDOUT` over global var `%stdout` meant that when
testing for output via stdout, we were not able to capture it through
rspec.

I assume this is because by setting @stream as `STDOUT`, we call
`STDOUT.puts(message)`, while puts is in fact syntax sugar for
`$stdout.puts(whatever)`. In any case, this fixes the issue with
capturing stdout for testing while not adversely affecting behaviour.